Error Handling.

+++ Day 29 & 30/main.py
@@ -43,27 +43,31 @@ def save():
     if len(website) == 0 or len(password) == 0:
         messagebox.showinfo(title="Oops", message="Make sure you haven't left any fields empty.")
     else:
-        # yes = messagebox.askyesno(title=website, message=f"These are the entered details: \n\nEmail: {email} \nPassword: {password} \n\nAre you sure you want to save this?")
+        yes = messagebox.askyesno(title=website, message=f"These are the entered details: \n\nEmail: {email} \nPassword: {password} \n\nAre you sure you want to save this?")
 
+        if yes:
+            try:
+                with open("data.json", mode="r") as data_file:
+                    # Read old data
+                    data = json.load(data_file)
 
-        with open("data.json", mode="r") as data_file:
-            # Read old data
-            data = json.load(data_file)
+            except FileNotFoundError:
+                with  open("data.json", mode="w") as data_file:
+                    json.dump(new_data, data_file, indent=4)
 
-            # Update old data
-            data.update(new_data)
+            else:
+                # Update old data
+                data.update(new_data)
 
+                with open("data.json", mode="w") as data_file:
+                    # Saving the updated data
+                    json.dump(data, data_file, indent=4)
 
-        with open("data.json", mode="w") as data_file:
-            # Saving the updated data
-            json.dump(data, data_file, indent=4)
+            finally:
+                website_entry.delete(0, END)
+                password_entry.delete(0, END)
 
-
-        if TRUE:
-            website_entry.delete(0, END)
-            password_entry.delete(0, END)
-
-            website_entry.focus()
+                website_entry.focus()
 
 
 # ---------------------------- UI SETUP ------------------------------- #
